requirements.txt
tensorflow-cpu==2.15.0 
keras==2.15.0 
numpy==1.26.4 
pandas==2.1.4 
fastapi 
uvicorn